Support a minted-input test case
Namely, this includes a new configuration parameter that can be
customized: minted_inputs

Defining `minted_inputs` sets all input blinding factors to
all-zeros. This is both helpful to test that the balancing still
works as expected, and we also test another invariant:
if all input blinding factors are 0 and there are two outputs,
the second output blinding factor should be the scalar negation of
the first.

Signed-off-by: Sam Stuewe <samuel.stuewe@gmail.com>